refactor: put location back into models folder
