+++ app/models/event.rb
@@ -5,7 +5,7 @@ require 'json'
 require 'sequel'
 require 'sequel/plugins/serialization'
 
-require_relative '../objects/location'
+require_relative 'location'
 
 Sequel::Plugins::Serialization.register_format(
   :location_s,
