+++ app/models/location.rb
