+++ app/models/postit.rb
@@ -4,7 +4,7 @@ require 'json'
 require 'sequel'
 require 'sequel/plugins/serialization'
 
-require_relative '../objects/location'
+require_relative 'location'
 
 Sequel::Plugins::Serialization.register_format(
   :location_s,
